extract Dropdowns
